v13: make AggregateProofType optional on ProveCommitSectors3Params & ProveReplicaUpdates3Params
Ref: https://github.com/filecoin-project/FIPs/pull/934

+++ abi/cbor_bytes.go
@@ -11,7 +11,7 @@ import (
 type CborBytes []byte
 
 func (t *CborBytes) MarshalCBOR(w io.Writer) error {
-	if len(*t) > cbg.ByteArrayMaxLen {
+	if uint64(len(*t)) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("byte array was too long")
 	}
