+++ builtin/v10/miner/cbor_gen.go
@@ -363,7 +363,7 @@ func (t *MinerInfo) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.ControlAddresses ([]address.Address) (slice)
-	if len(t.ControlAddresses) > cbg.MaxLength {
+	if uint64(len(t.ControlAddresses)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.ControlAddresses was too long")
 	}
 
@@ -383,7 +383,7 @@ func (t *MinerInfo) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.PeerId ([]uint8) (slice)
-	if len(t.PeerId) > cbg.ByteArrayMaxLen {
+	if uint64(len(t.PeerId)) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("Byte array in field t.PeerId was too long")
 	}
 
@@ -396,7 +396,7 @@ func (t *MinerInfo) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Multiaddrs ([][]uint8) (slice)
-	if len(t.Multiaddrs) > cbg.MaxLength {
+	if uint64(len(t.Multiaddrs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Multiaddrs was too long")
 	}
 
@@ -404,7 +404,7 @@ func (t *MinerInfo) MarshalCBOR(w io.Writer) error {
 		return err
 	}
 	for _, v := range t.Multiaddrs {
-		if len(v) > cbg.ByteArrayMaxLen {
+		if uint64(len(v)) > cbg.ByteArrayMaxLen {
 			return xerrors.Errorf("Byte array in field v was too long")
 		}
 
@@ -796,7 +796,7 @@ func (t *Deadlines) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Due ([48]cid.Cid) (array)
-	if len(t.Due) > cbg.MaxLength {
+	if uint64(len(t.Due)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Due was too long")
 	}
 
@@ -1533,6 +1533,7 @@ func (t *SectorPreCommitOnChainInfo) MarshalCBOR(w io.Writer) error {
 			return err
 		}
 	}
+
 	return nil
 }
 
@@ -1654,7 +1655,7 @@ func (t *SectorPreCommitInfo) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.DealIDs ([]abi.DealID) (slice)
-	if len(t.DealIDs) > cbg.MaxLength {
+	if uint64(len(t.DealIDs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.DealIDs was too long")
 	}
 
@@ -1925,7 +1926,7 @@ func (t *SectorOnChainInfo) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.DealIDs ([]abi.DealID) (slice)
-	if len(t.DealIDs) > cbg.MaxLength {
+	if uint64(len(t.DealIDs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.DealIDs was too long")
 	}
 
@@ -2339,6 +2340,7 @@ func (t *WorkerKeyChange) MarshalCBOR(w io.Writer) error {
 			return err
 		}
 	}
+
 	return nil
 }
 
@@ -2417,7 +2419,7 @@ func (t *VestingFunds) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Funds ([]miner.VestingFund) (slice)
-	if len(t.Funds) > cbg.MaxLength {
+	if uint64(len(t.Funds)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Funds was too long")
 	}
 
@@ -2609,7 +2611,7 @@ func (t *WindowedPoSt) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Proofs ([]proof.PoStProof) (slice)
-	if len(t.Proofs) > cbg.MaxLength {
+	if uint64(len(t.Proofs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Proofs was too long")
 	}
 
@@ -2802,6 +2804,7 @@ func (t *BeneficiaryTerm) MarshalCBOR(w io.Writer) error {
 			return err
 		}
 	}
+
 	return nil
 }
 
@@ -3049,7 +3052,7 @@ func (t *GetControlAddressesReturn) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.ControlAddrs ([]address.Address) (slice)
-	if len(t.ControlAddrs) > cbg.MaxLength {
+	if uint64(len(t.ControlAddrs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.ControlAddrs was too long")
 	}
 
@@ -3167,7 +3170,7 @@ func (t *ChangeWorkerAddressParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.NewControlAddrs ([]address.Address) (slice)
-	if len(t.NewControlAddrs) > cbg.MaxLength {
+	if uint64(len(t.NewControlAddrs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.NewControlAddrs was too long")
 	}
 
@@ -3271,7 +3274,7 @@ func (t *ChangePeerIDParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.NewID ([]uint8) (slice)
-	if len(t.NewID) > cbg.ByteArrayMaxLen {
+	if uint64(len(t.NewID)) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("Byte array in field t.NewID was too long")
 	}
 
@@ -3355,7 +3358,7 @@ func (t *SubmitWindowedPoStParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Partitions ([]miner.PoStPartition) (slice)
-	if len(t.Partitions) > cbg.MaxLength {
+	if uint64(len(t.Partitions)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Partitions was too long")
 	}
 
@@ -3370,7 +3373,7 @@ func (t *SubmitWindowedPoStParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Proofs ([]proof.PoStProof) (slice)
-	if len(t.Proofs) > cbg.MaxLength {
+	if uint64(len(t.Proofs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Proofs was too long")
 	}
 
@@ -3396,7 +3399,7 @@ func (t *SubmitWindowedPoStParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.ChainCommitRand (abi.Randomness) (slice)
-	if len(t.ChainCommitRand) > cbg.ByteArrayMaxLen {
+	if uint64(len(t.ChainCommitRand)) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("Byte array in field t.ChainCommitRand was too long")
 	}
 
@@ -3623,7 +3626,7 @@ func (t *PreCommitSectorParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.DealIDs ([]abi.DealID) (slice)
-	if len(t.DealIDs) > cbg.MaxLength {
+	if uint64(len(t.DealIDs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.DealIDs was too long")
 	}
 
@@ -3925,7 +3928,7 @@ func (t *ProveCommitSectorParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Proof ([]uint8) (slice)
-	if len(t.Proof) > cbg.ByteArrayMaxLen {
+	if uint64(len(t.Proof)) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("Byte array in field t.Proof was too long")
 	}
 
@@ -4017,7 +4020,7 @@ func (t *ExtendSectorExpirationParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Extensions ([]miner.ExpirationExtension) (slice)
-	if len(t.Extensions) > cbg.MaxLength {
+	if uint64(len(t.Extensions)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Extensions was too long")
 	}
 
@@ -4112,7 +4115,7 @@ func (t *ExtendSectorExpiration2Params) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Extensions ([]miner.ExpirationExtension2) (slice)
-	if len(t.Extensions) > cbg.MaxLength {
+	if uint64(len(t.Extensions)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Extensions was too long")
 	}
 
@@ -4207,7 +4210,7 @@ func (t *TerminateSectorsParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Terminations ([]miner.TerminationDeclaration) (slice)
-	if len(t.Terminations) > cbg.MaxLength {
+	if uint64(len(t.Terminations)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Terminations was too long")
 	}
 
@@ -4366,7 +4369,7 @@ func (t *DeclareFaultsParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Faults ([]miner.FaultDeclaration) (slice)
-	if len(t.Faults) > cbg.MaxLength {
+	if uint64(len(t.Faults)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Faults was too long")
 	}
 
@@ -4461,7 +4464,7 @@ func (t *DeclareFaultsRecoveredParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Recoveries ([]miner.RecoveryDeclaration) (slice)
-	if len(t.Recoveries) > cbg.MaxLength {
+	if uint64(len(t.Recoveries)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Recoveries was too long")
 	}
 
@@ -4556,7 +4559,7 @@ func (t *DeferredCronEventParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.EventPayload ([]uint8) (slice)
-	if len(t.EventPayload) > cbg.ByteArrayMaxLen {
+	if uint64(len(t.EventPayload)) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("Byte array in field t.EventPayload was too long")
 	}
 
@@ -4794,7 +4797,7 @@ func (t *ReportConsensusFaultParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.BlockHeader1 ([]uint8) (slice)
-	if len(t.BlockHeader1) > cbg.ByteArrayMaxLen {
+	if uint64(len(t.BlockHeader1)) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("Byte array in field t.BlockHeader1 was too long")
 	}
 
@@ -4807,7 +4810,7 @@ func (t *ReportConsensusFaultParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.BlockHeader2 ([]uint8) (slice)
-	if len(t.BlockHeader2) > cbg.ByteArrayMaxLen {
+	if uint64(len(t.BlockHeader2)) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("Byte array in field t.BlockHeader2 was too long")
 	}
 
@@ -4820,7 +4823,7 @@ func (t *ReportConsensusFaultParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.BlockHeaderExtra ([]uint8) (slice)
-	if len(t.BlockHeaderExtra) > cbg.ByteArrayMaxLen {
+	if uint64(len(t.BlockHeaderExtra)) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("Byte array in field t.BlockHeaderExtra was too long")
 	}
 
@@ -4998,7 +5001,7 @@ func (t *ConfirmSectorProofsParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Sectors ([]abi.SectorNumber) (slice)
-	if len(t.Sectors) > cbg.MaxLength {
+	if uint64(len(t.Sectors)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Sectors was too long")
 	}
 
@@ -5141,7 +5144,7 @@ func (t *ChangeMultiaddrsParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.NewMultiaddrs ([][]uint8) (slice)
-	if len(t.NewMultiaddrs) > cbg.MaxLength {
+	if uint64(len(t.NewMultiaddrs)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.NewMultiaddrs was too long")
 	}
 
@@ -5149,7 +5152,7 @@ func (t *ChangeMultiaddrsParams) MarshalCBOR(w io.Writer) error {
 		return err
 	}
 	for _, v := range t.NewMultiaddrs {
-		if len(v) > cbg.ByteArrayMaxLen {
+		if uint64(len(v)) > cbg.ByteArrayMaxLen {
 			return xerrors.Errorf("Byte array in field v was too long")
 		}
 
@@ -5471,7 +5474,7 @@ func (t *PreCommitSectorBatchParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Sectors ([]miner.PreCommitSectorParams) (slice)
-	if len(t.Sectors) > cbg.MaxLength {
+	if uint64(len(t.Sectors)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Sectors was too long")
 	}
 
@@ -5571,7 +5574,7 @@ func (t *ProveCommitAggregateParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.AggregateProof ([]uint8) (slice)
-	if len(t.AggregateProof) > cbg.ByteArrayMaxLen {
+	if uint64(len(t.AggregateProof)) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("Byte array in field t.AggregateProof was too long")
 	}
 
@@ -5658,7 +5661,7 @@ func (t *ProveReplicaUpdatesParams) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Updates ([]miner.ReplicaUpdate) (slice)
-	if len(t.Updates) > cbg.MaxLength {
+	if uint64(len(t.Updates)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Updates was too long")
 	}
 
@@ -5762,6 +5765,7 @@ func (t *CronEventPayload) MarshalCBOR(w io.Writer) error {
 			return err
 		}
 	}
+
 	return nil
 }
 
@@ -5831,7 +5835,7 @@ func (t *PreCommitSectorBatchParams2) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Sectors ([]miner.SectorPreCommitInfo) (slice)
-	if len(t.Sectors) > cbg.MaxLength {
+	if uint64(len(t.Sectors)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Sectors was too long")
 	}
 
@@ -5926,7 +5930,7 @@ func (t *ProveReplicaUpdatesParams2) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Updates ([]miner.ReplicaUpdate2) (slice)
-	if len(t.Updates) > cbg.MaxLength {
+	if uint64(len(t.Updates)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Updates was too long")
 	}
 
@@ -6040,6 +6044,7 @@ func (t *ChangeBeneficiaryParams) MarshalCBOR(w io.Writer) error {
 			return err
 		}
 	}
+
 	return nil
 }
 
@@ -6287,7 +6292,7 @@ func (t *GetPeerIDReturn) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.PeerId ([]uint8) (slice)
-	if len(t.PeerId) > cbg.ByteArrayMaxLen {
+	if uint64(len(t.PeerId)) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("Byte array in field t.PeerId was too long")
 	}
 
@@ -6365,7 +6370,7 @@ func (t *GetMultiAddrsReturn) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.MultiAddrs ([]uint8) (slice)
-	if len(t.MultiAddrs) > cbg.ByteArrayMaxLen {
+	if uint64(len(t.MultiAddrs)) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("Byte array in field t.MultiAddrs was too long")
 	}
 
@@ -6661,6 +6666,7 @@ func (t *ExpirationExtension) MarshalCBOR(w io.Writer) error {
 			return err
 		}
 	}
+
 	return nil
 }
 
@@ -6963,7 +6969,7 @@ func (t *ReplicaUpdate) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Deals ([]abi.DealID) (slice)
-	if len(t.Deals) > cbg.MaxLength {
+	if uint64(len(t.Deals)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Deals was too long")
 	}
 
@@ -6990,7 +6996,7 @@ func (t *ReplicaUpdate) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.ReplicaProof ([]uint8) (slice)
-	if len(t.ReplicaProof) > cbg.ByteArrayMaxLen {
+	if uint64(len(t.ReplicaProof)) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("Byte array in field t.ReplicaProof was too long")
 	}
 
@@ -7220,7 +7226,7 @@ func (t *ReplicaUpdate2) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.Deals ([]abi.DealID) (slice)
-	if len(t.Deals) > cbg.MaxLength {
+	if uint64(len(t.Deals)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.Deals was too long")
 	}
 
@@ -7247,7 +7253,7 @@ func (t *ReplicaUpdate2) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.ReplicaProof ([]uint8) (slice)
-	if len(t.ReplicaProof) > cbg.ByteArrayMaxLen {
+	if uint64(len(t.ReplicaProof)) > cbg.ByteArrayMaxLen {
 		return xerrors.Errorf("Byte array in field t.ReplicaProof was too long")
 	}
 
@@ -7476,7 +7482,7 @@ func (t *ExpirationExtension2) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.SectorsWithClaims ([]miner.SectorClaim) (slice)
-	if len(t.SectorsWithClaims) > cbg.MaxLength {
+	if uint64(len(t.SectorsWithClaims)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.SectorsWithClaims was too long")
 	}
 
@@ -7500,6 +7506,7 @@ func (t *ExpirationExtension2) MarshalCBOR(w io.Writer) error {
 			return err
 		}
 	}
+
 	return nil
 }
 
@@ -7650,7 +7657,7 @@ func (t *SectorClaim) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.MaintainClaims ([]verifreg.ClaimId) (slice)
-	if len(t.MaintainClaims) > cbg.MaxLength {
+	if uint64(len(t.MaintainClaims)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.MaintainClaims was too long")
 	}
 
@@ -7666,7 +7673,7 @@ func (t *SectorClaim) MarshalCBOR(w io.Writer) error {
 	}
 
 	// t.DropClaims ([]verifreg.ClaimId) (slice)
-	if len(t.DropClaims) > cbg.MaxLength {
+	if uint64(len(t.DropClaims)) > cbg.MaxLength {
 		return xerrors.Errorf("Slice value in field t.DropClaims was too long")
 	}
